move package_config dep to its own library so it doesn\'t appear in the test isolate deps (#1243)
